Moved model repository and mysql outside the user-api folder
