+++ internal/app/repository/user_repository.go
@@ -3,8 +3,8 @@ package repository
 import (
 	"errors"
 
+	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/model"
 	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/user-api/constant"
-	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/user-api/model"
 	"github.com/google/uuid"
 	"gorm.io/gorm"
 )
