+++ internal/app/repository/user_repository_test.go
@@ -4,8 +4,8 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/user-api/model"
-	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/user-api/repository"
+	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/model"
+	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/repository"
 	"github.com/google/uuid"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
