+++ internal/app/user-api/service/user_service.go
@@ -5,10 +5,10 @@ import (
 	"net/http"
 	"strings"
 
+	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/model"
+	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/repository"
 	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/user-api/constant"
 	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/user-api/errormessage"
-	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/user-api/model"
-	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/user-api/repository"
 	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/user-api/validation"
 	"github.com/google/uuid"
 )
