+++ internal/app/user-api/validation/user_validator_test.go
@@ -3,9 +3,9 @@ package validation
 import (
 	"testing"
 
+	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/model"
+	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/repository"
 	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/user-api/constant"
-	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/user-api/model"
-	"github.com/afernandowex/ps-tag-onboarding-go/internal/app/user-api/repository"
 	"gorm.io/driver/sqlite"
 	"gorm.io/gorm"
 )
